feat!: move `HierarchicalExpression` into place of `TemplateExpression`
