+++ examples/template_expression.jl
@@ -6,20 +6,27 @@ using Test: @test
 options = Options(; binary_operators=(+, *, /, -), unary_operators=(sin, cos))
 operators = options.operators
 variable_names = (i -> "x$i").(1:3)
-x1, x2, x3 = (i -> Expression(Node(Float64; feature=i); operators, variable_names)).(1:3)
+x1, x2, x3 = (i -> ComposableExpression(Node(Float64; feature=i); operators, variable_names)).(1:3)
 
-structure = TemplateStructure{(:f, :g1, :g2)}(;
-    combine_vectors=e -> map((f, g1, g2) -> (f + g1, f + g2), e.f, e.g1, e.g2),
-    combine_strings=e -> "( $(e.f) + $(e.g1), $(e.f) + $(e.g2) )",
-    variable_constraints=(; f=[1, 2], g1=[3], g2=[3]),
+structure = TemplateStructure{(:f, :g1, :g2)}(
+    ((; f, g1, g2), (x1, x2, x3)) -> let
+        _f = f(x1, x2)
+        _g1 = g1(x3)
+        _g2 = g2(x3)
+        _out1 = _f + _g1
+        _out2 = _f + _g2
+        ValidVector(map(tuple, _out1.x, _out2.x), _out1.valid && _out2.valid)
+    end,
 )
 
 st_expr = TemplateExpression((; f=x1, g1=x3, g2=x3); structure, operators, variable_names)
 
-X = rand(100, 3) .* 10
+x1 = rand(100)
+x2 = rand(100)
+x3 = rand(100)
 
 # Our dataset is a vector of 2-tuples
-y = [(sin(X[i, 1]) + X[i, 3]^2, sin(X[i, 1]) + X[i, 3]) for i in eachindex(axes(X, 1))]
+y = [(sin(x1[i]) + x3[i]^2, sin(x1[i]) + x3[i]) for i in eachindex(x1, x2, x3)]
 
 model = SRRegressor(;
     binary_operators=(+, *),
@@ -32,7 +39,7 @@ model = SRRegressor(;
     early_stop_condition=(loss, complexity) -> loss < 1e-5 && complexity <= 7,
 )
 
-mach = machine(model, X, y)
+mach = machine(model, [x1 x2 x3], y)
 fit!(mach)
 
 # Check the performance of the model
@@ -48,6 +55,6 @@ best_f = get_contents(best_expr).f
 best_g1 = get_contents(best_expr).g1
 best_g2 = get_contents(best_expr).g2
 
-@test best_f(X') ≈ (@. sin(X[:, 1]))
-@test best_g1(X') ≈ (@. X[:, 3] * X[:, 3])
-@test best_g2(X') ≈ (@. X[:, 3])
+@test best_f(x1, x2) ≈ @. sin.(x1)
+@test best_g1(x3) ≈ (@. x3 * x3)
+@test best_g2(x3) ≈ (@. x3)
